Split "tests/test_add_options.py" into multiple files
